fix: enable_dev_tools
