+++ examples/counter/src/lib.rs
@@ -56,11 +56,5 @@ fn view(model: &Model) -> Node<Msg> {
 #[wasm_bindgen(start)]
 pub fn start() {
     // App::start("app", init, update, view);
-
-    // // Approach with options works
-    // App::start_with_options("app", init, update, view, vec![Box::new(WithDevtools::new())]);
-    
-    // // This one does not
     App::start("app", init, update, view).enable_dev_tools();
-
 }
